fix: Use registry pull secret when building images
Updates `getBuildConfig` in the OpenShift client to use the registry
secret to pull the base Docker image.

When OpenShift is used to create new Docker images, a "BuildConfig" is
created. This build config provides OpensShift with instructions on how
to build the Docker image. The current build config was using a secret
to push the build image into a registry, but it was not using this
secret when pulling the base Docker image. This means that it only
worked for base images in registries that allow anonymous access. This
commit updates the BuildConfig to use the registry secret when pulling
the image.